Move helper functions to _utils.
