+++ nilearn/mass_univariate/permuted_least_squares.py
@@ -11,142 +11,16 @@ import joblib
 import nibabel as nib
 import numpy as np
 from nilearn.masking import apply_mask
-from scipy import linalg, ndimage, stats
+from scipy import ndimage, stats
 from sklearn.utils import check_random_state
 
-from ._utils import _calculate_cluster_measures, _null_to_p
-
-
-def _normalize_matrix_on_axis(m, axis=0):
-    """ Normalize a 2D matrix on an axis.
-
-    Parameters
-    ----------
-    m : numpy 2D array,
-        The matrix to normalize.
-
-    axis : integer in {0, 1}, optional
-        A valid axis to normalize across.
-        Default=0.
-
-    Returns
-    -------
-    ret : numpy array, shape = m.shape
-        The normalized matrix
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from nilearn.mass_univariate.permuted_least_squares import (
-    ...     _normalize_matrix_on_axis)
-    >>> X = np.array([[0, 4], [1, 0]])
-    >>> _normalize_matrix_on_axis(X)
-    array([[0., 1.],
-           [1., 0.]])
-    >>> _normalize_matrix_on_axis(X, axis=1)
-    array([[0., 1.],
-           [1., 0.]])
-
-    """
-    if m.ndim > 2:
-        raise ValueError('This function only accepts 2D arrays. '
-                         'An array of shape %r was passed.' % m.shape)
-
-    if axis == 0:
-        # array transposition preserves the contiguity flag of that array
-        ret = (m.T / np.sqrt(np.sum(m ** 2, axis=0))[:, np.newaxis]).T
-    elif axis == 1:
-        ret = _normalize_matrix_on_axis(m.T).T
-    else:
-        raise ValueError('axis(=%d) out of bounds' % axis)
-    return ret
-
-
-def _orthonormalize_matrix(m, tol=1.e-12):
-    """ Orthonormalize a matrix.
-
-    Uses a Singular Value Decomposition.
-    If the input matrix is rank-deficient, then its shape is cropped.
-
-    Parameters
-    ----------
-    m : numpy array,
-        The matrix to orthonormalize.
-
-    tol: float, optional
-        Tolerance parameter for nullity. Default=1e-12.
-
-    Returns
-    -------
-    ret : numpy array, shape = m.shape
-        The orthonormalized matrix.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from nilearn.mass_univariate.permuted_least_squares import (
-    ...     _orthonormalize_matrix)
-    >>> X = np.array([[1, 2], [0, 1], [1, 1]])
-    >>> _orthonormalize_matrix(X)
-    array([[-0.81049889, -0.0987837 ],
-           [-0.31970025, -0.75130448],
-           [-0.49079864,  0.65252078]])
-    >>> X = np.array([[0, 1], [4, 0]])
-    >>> _orthonormalize_matrix(X)
-    array([[ 0., -1.],
-           [-1.,  0.]])
-
-    """
-    U, s, _ = linalg.svd(m, full_matrices=False)
-    n_eig = np.count_nonzero(s > tol)
-    return np.ascontiguousarray(U[:, :n_eig])
-
-
-def _t_score_with_covars_and_normalized_design(tested_vars, target_vars,
-                                               covars_orthonormalized=None):
-    """t-score in the regression of tested variates against target variates
-
-    Covariates are taken into account (if not None).
-    The normalized_design case corresponds to the following assumptions:
-    - tested_vars and target_vars are normalized
-    - covars_orthonormalized are orthonormalized
-    - tested_vars and covars_orthonormalized are orthogonal
-      (np.dot(tested_vars.T, covars) == 0)
-
-    Parameters
-    ----------
-    tested_vars : array-like, shape=(n_samples, n_tested_vars)
-        Explanatory variates.
-
-    target_vars : array-like, shape=(n_samples, n_target_vars)
-        Targets variates. F-ordered is better for efficient computation.
-
-    covars_orthonormalized : array-like, shape=(n_samples, n_covars) or None, \
-            optional
-        Confounding variates.
-
-    Returns
-    -------
-    score : numpy.ndarray, shape=(n_target_vars, n_tested_vars)
-        t-scores associated with the tests of each explanatory variate against
-        each target variate (in the presence of covars).
-
-    """
-    if covars_orthonormalized is None:
-        lost_dof = 0
-    else:
-        lost_dof = covars_orthonormalized.shape[1]
-    # Tested variates are fitted independently,
-    # so lost_dof is unrelated to n_tested_vars.
-    dof = target_vars.shape[0] - lost_dof
-    beta_targetvars_testedvars = np.dot(target_vars.T, tested_vars)
-    if covars_orthonormalized is None:
-        rss = (1 - beta_targetvars_testedvars ** 2)
-    else:
-        beta_targetvars_covars = np.dot(target_vars.T, covars_orthonormalized)
-        a2 = np.sum(beta_targetvars_covars ** 2, 1)
-        rss = (1 - a2[:, np.newaxis] - beta_targetvars_testedvars ** 2)
-    return beta_targetvars_testedvars * np.sqrt((dof - 1.) / rss)
+from nilearn.mass_univariate._utils import (
+    _calculate_cluster_measures,
+    _normalize_matrix_on_axis,
+    _null_to_p,
+    _orthonormalize_matrix,
+    _t_score_with_covars_and_normalized_design,
+)
 
 
 def _permuted_ols_on_chunk(
